SCLang: Refactor primitives and keyword arguments
From now on, primitives in c++ and sclang must have the same number of arguments, including variable arguments and variable keyword arguments.

Previously, primitives could be declared with `n` arguments in c++ but `m` arguments in sclang. This was very confusing.

Previously, having a primitive with keyword arguments meant all keyword arguments were passed with the key to the primitive, even if they were declared as positional in sclang. This meant that
```
A {
   *new { |foo... args, kwargs|
          _PrintAllArgs
   }
}
A(foo: 10, bar: 20)
```
would print `[\foo, 10, \bar, 20]` rather than `[10, \bar, 20]`, which is unique behaviour.

Fixes bug `().put(\asdf, 123, meow: 23)` which would crash because the stack pointer wasn't decremented when removing them from the stack. I do not know when this bug was introduced.

Introduces a number of tests to ensure primitives will always be working correctly.

Refactors compiling of block as the logic was very complicated with lots of mutable variables.