sclang: non recursive method macro of double dispatch
The existing way of doing double dispatch through the performBinaryOp* methods doesn't work because they don't actually 'perform', likewise, in sc, you can't know if something will perform ahead of time, as doesNotUnderstand hasn't been called.

This is very important as we move towards AbstractObject as being able to call a perform* method and it actually perform, going through doesNotUnderstand is imperative.

Implements a special kind of method that will throw an error if recursion is detected. This simply allows us to flip the arguments in binary operators and try them without getting stuck.