classlib: limit use of `respondsTo` and post warning
`respondsTo` is bad smalltalk design as it doesn't allow of `doesNotUnderstand` to take effect. While this is currently a minor issue in the class library, as we move towards AbtractObject, this will become a major pain.

This PR removes the vast majority of uses from the class library as they are trivially replaceable with `tryPerform`, the remaining places require dependency inversion to solve, but can be solved simply with flags.

Also makes `respondsTo` post a warning when used, this can be disabled.