coc: no ai slop
Introduce a 'no ai slop' section to the code of conduct.
This text is (slightly) adapted from luka's response on the forum, which is in turn a adaptation of @capital-G's proposal.

https://scsynth.org/t/add-no-ai-slop-to-community-guidelines/12672/18?u=jordan

@jamshark70 has suggested the following procedure:

1. This PR be reviewed and edited until a consensus is reached.
2. This then be shared over on the forum (and where ever else).
3. Only then should this be merged.

Please avoid turning this into a discussion on AI in general (perhaps open a new issue), this is about halting the proliferation of 'low effort'/'ai slop' throughout the community.